requirements.txt
# requirements.txt
instagrapi
python-dotenv
rich